fix startup script

+++ ib_manifest_util/update_repository.py
@@ -25,6 +25,7 @@ def update_repo(
     output_hardening_path: str | Path | None = None,
     output_dockerfile_path: str | Path | None = None,
     dockerfile_template_path: str | Path = None,
+    cleanup: bool = False,
 ):
     """High level function to update an Iron Bank repository with a new environment.
 
@@ -143,9 +144,10 @@ def update_repo(
         dockerfile_template_path=dockerfile_template_path,
     )
 
-    # clean up ib_manifest.yaml
-    logger.info(
-        f"Cleaning up. Removing {ib_manifest_path} and local conda channel {local_channel_path}"
-    )
-    ib_manifest_path.unlink()
-    shutil.rmtree(local_channel_path)
+    if cleanup:
+        # clean up ib_manifest.yaml
+        logger.info(
+            f"Cleaning up. Removing {ib_manifest_path} and local conda channel {local_channel_path}"
+        )
+        ib_manifest_path.unlink()
+        shutil.rmtree(local_channel_path)
